Move minio client factory to AbstractDatasetView.
Moving factory class method within DatasetApiView for creating a MinIO
client object up a level to abstract superclass, to facilitate reuse.
